First commit. One interface, one implementing class. Non-compilable state.
TODOS:

(1) Switch from implementing classes to defining their API and authoring the testcases. The testcases need to be common. Only thing you should be changing is the object pointed to by the Bag reference.

(2)  Create a client program that will time the access of the elements after shaking the bag. Goal: estimate impact of partial cache storage of large arrays.